+++ src/main/java/org/example/App.java
@@ -1,13 +1,12 @@
 package org.example;
 
-import org.example.ManualConstructor.persistence.DataBasePersistence;
-import org.example.ManualConstructor.persistence.FilePersistence;
-import org.example.ManualConstructor.persistence.PersistenceLayer;
-import org.example.ManualConstructor.service.AdvancedService;
+import org.example.container.Container;
+import org.example.persistence.DataBasePersistence;
+import org.example.persistence.FilePersistence;
+import org.example.persistence.PersistenceLayer;
+import org.example.service.AdvancedService;
 import org.jboss.weld.environment.se.Weld;
 import org.jboss.weld.environment.se.WeldContainer;
-
-import java.lang.reflect.Constructor;
 
 
 public class App {
@@ -35,7 +34,9 @@ public class App {
         Weld weld = new Weld();
         try (WeldContainer weldContainer = weld.initialize()) {
             AdvancedService service = weldContainer.select(AdvancedService.class).get();
+            Car car2 = weldContainer.select(Car.class).get();
 
+            System.out.println(car2.toString() + " has been created.");
             service.save("TestData1");
             service.update("TestData2");
             service.delete("OldData");
