+++ src/main/java/org/example/service/Service.java
@@ -1,4 +1,4 @@
-package org.example.ManualConstructor.service;
+package org.example.service;
 
 public interface Service {
     void processData(String data);
